requirements.txt
python-telegram-bot==20.0
python-telegram-bot[job-queue]
openai
requests
bs4
python-dotenv
aiofiles>=23.1.0
setuptools<81
selenium